moved Today view into Activity and did some cleanup
