+++ src/views/Activity.vue
@@ -7,17 +7,6 @@ div
   b-alert(variant="danger" :show="errormsg.length > 0")
     | {{ errormsg }}
 
-  b-alert(show)
-    | Check out the new today view! Currently in alpha.
-    b-button(style="float: right" :to="'/today/' + host" size="sm" variant="info") Open today view
-
-  //div.alert.alert-info(role="alert")
-    small
-      | This page is currently considered early work and is known to have issues:
-      li
-        a(href='https://github.com/ActivityWatch/aw-webui/issues/22')
-          | Needs a date selector
-
   b-button-group
     b-button(:to="'/activity/' + host + '/' + previousDay()")
       icon(name="arrow-left")
@@ -25,7 +14,7 @@ div
     b-button(:to="'/activity/' + host + '/' + nextDay()", :disabled="nextDay() > today")
       |  Next day
       icon(name="arrow-right")
-  b-button(v-on:click="query()", style="margin-left: 1rem;")
+  b-button(v-on:click="refresh()", style="margin-left: 1rem;")
     icon(name="refresh")
     |  Refresh
 
@@ -75,6 +64,16 @@ div
 
   div#apptimeline-container
 
+  hr
+
+  div
+    h4 Clock
+
+    b-alert(variant="warning" show)
+      | #[b Note:] This is an early version. It has known issues that will be resolved in a future update.
+      | See #[a(href="https://github.com/ActivityWatch/aw-webui/issues/36") issue #36] for details.
+
+    aw-sunburst(:hierarchy="hierarchy")
 </template>
 
 <style lang="scss">
@@ -100,9 +99,12 @@ import 'vue-awesome/icons/arrow-left'
 import 'vue-awesome/icons/arrow-right'
 import 'vue-awesome/icons/refresh'
 
+import Sunburst from '../visualizations/Sunburst.vue';
+
 let $QueryView  = Resources.$QueryView;
 let $CreateView  = Resources.$CreateView;
 let $Info  = Resources.$Info;
+let $Bucket = Resources.$Bucket;
 let $Event  = Resources.$Event;
 
 var daylength = 86400000;
@@ -120,21 +122,28 @@ export default {
       duration: "",
       eventcount: 0,
       errormsg: "",
-      numberOfWindowTitles: 5
+      numberOfWindowTitles: 5,
+
+      // Sunburst stuff
+      hierarchy: Object,
     }
+  },
+
+  components: {
+    "aw-sunburst": Sunburst,
   },
 
   watch: {
     '$route': function(to, from) {
       console.log("Route changed")
-      this.query();
+      this.refresh();
     },
     'filterAFK': function(to, from) {
-      this.query();
+      this.refresh();
     },
     'timelineShowAFK': function(to, from) {
-      this.query();
-    }
+      this.refresh();
+    },
   },
 
   computed: {
@@ -152,12 +161,17 @@ export default {
     summary.create(document.getElementById("windowtitles-container"));
     timeline.create(document.getElementById("apptimeline-container"));
 
-    this.query();
+    this.refresh();
   },
 
   methods: {
     previousDay: function() { return moment(this.dateStart).subtract(1, 'days').format("YYYY-MM-DD") },
     nextDay: function() { return moment(this.dateStart).add(1, 'days').format("YYYY-MM-DD") },
+
+    refresh: function() {
+      this.query();
+      this.visualize();
+    },
 
     errorHandler: function(response) {
       console.error(response);
@@ -189,16 +203,16 @@ export default {
 
     todaysEvents: function(bucket_id) {
       let today = moment(this.dateStart);
+      let endOfToday = moment(today).add(1, "days");
       return $Event.get({id: bucket_id, limit: -1,
-                         start: today.format(), end: today.add(1, "days").format()});
+                         start: today.format(), end: endOfToday.format()})
+                   .then((response) => response.json());
     },
 
     windowEventsFilteredByAFK: function() {
       return this.todaysEvents(this.windowBucketId)
-        .then((response) => {
-          let events = response.json();
-          return this.todaysEvents(this.afkBucketId).then((response) => {
-              let afkevents = response.json();
+        .then((events) => {
+          return this.todaysEvents(this.afkBucketId).then((afkevents) => {
               if(this.filterAFK) {
                 events = event_parsing.filterAFKTime(events, afkevents);
               }
@@ -252,10 +266,11 @@ export default {
       timeline.set_status(timeline_elem, "Loading...");
 
       let today = moment(this.dateStart);
+      let endOfToday = today.clone().add(1, 'days');
       $QueryView.get({"viewname": viewname,
                       "limit": -1,
                       "start": today.format(),
-                      "end": today.add(1, 'days').format()})
+                      "end": endOfToday.format()})
         .then((response) => {
           if (response.status > 304){
             this.errorHandler(response);
@@ -297,6 +312,138 @@ export default {
           }] : []
         }]
       };
+    },
+
+    // Everything below is related to the sunburst visualization
+    getBucketInfo: function(bucket_id) {
+      return $Bucket.get({"id": bucket_id}).then((response) => {
+        return response.json();
+      });
+    },
+
+    visualize: function() {
+      function buildHierarchy(parents, children) {
+          parents = _.sortBy(parents, "timestamp", "desc");
+          children = _.sortBy(children, "timestamp", "desc");
+
+          var i_child = 0;
+          for(var i_parent = 0; i_parent < parents.length; i_parent++) {
+              let p = parents[i_parent];
+              let p_start = moment(p.timestamp);
+              let p_end = p_start.clone().add(p.duration, "seconds");
+
+              p.children = [];
+              while(i_child < children.length) {
+                  var e = children[i_child];
+                  var e_start = moment(e.timestamp);
+                  var e_end = e_start.clone().add(e.duration, "seconds");
+
+                  let too_far = e_start.isAfter(p_end);
+                  let before_parent = e_end.isBefore(p_start);
+                  let within_parent = e_start.isAfter(p_start) && e_end.isBefore(p_end);
+                  let after_parent = e_start.isAfter(p_end);
+
+                  // TODO: This isn't correct, yet
+                  if(before_parent) {
+                    // Child is behind parent
+                    //console.log("Too far behind: " + i_child);
+                    i_child++;
+                  } else if(within_parent) {
+                    //console.log("Added relation: " + i_child);
+                    p.children = _.concat(p.children, e);
+                    i_child++;
+                  } else if(after_parent) {
+                    // Child is ahead of parent
+                    //console.log("Too far ahead: " + i_child);
+                    break;
+                  } else {
+                    // TODO: Split events when this happens
+                    console.warn("Between parents");
+                    p.children = _.concat(p.children, e);
+                    i_child++;
+                  }
+              }
+          }
+
+          // Build the root node
+          console.log(parents);
+          let m_start = moment(_.first(parents).timestamp)
+          let m_end = moment(_.tail(parents).timestamp)
+          let duration = (m_end - m_start) / 1000;
+          return {
+            "timestamp": _.first(parents).timestamp,
+            // TODO: If we want a 12/24h clock, this has to change
+            "duration": duration,
+            "data": {"title": "ROOT"},
+            "children": parents
+          }
+      }
+
+      function chunkHierarchy(events, key) {
+          // TODO: Merge window events with same app and assign the title events as children
+          let new_events = [events[0]];
+          let p_i = 0;
+          _.each(events, (e, i) => {
+              if(e.data[key] === new_events[p_i].data[key]) {
+                  //console.log("merge");
+                  let e_moment = moment(e.timestamp);
+                  let ne_moment = moment(new_events[p_i].timestamp);
+                  new_events[p_i].duration = -e_moment.diff(ne_moment, "seconds", true) + e.duration;
+                  console.log(new_events[p_i].duration);
+              } else {
+                  //console.log("skip");
+                  //console.log(new_events[p_i].duration);
+                  p_i++;
+                  new_events[p_i] = e;
+              }
+          });
+          _.each(new_events, (e, i) => {
+              // Get rid of other keys
+              e.data = _.pickBy(e.data, (v, k) => k === key);
+          })
+          return new_events;
+      }
+
+      function chunkHierarchy2(events, key) {
+        events = _.sortBy(events, (e) => e.timestamp);
+        events = _.reverse(events);
+        events = _.reduce(events,
+          function(acc, e) {
+            let last = _.last(acc);
+            if(last.data[key] === e.data[key]) {
+              last.duration = moment(e.timestamp).diff(last.timestamp, "seconds", true) + e.duration;
+            } else {
+              acc.push(e);
+            }
+            return acc;
+          },
+          [events[0]]);
+        return events;
+      }
+
+      this.todaysEvents(this.afkBucketId).then((events_afk) => {
+        this.todaysEvents(this.windowBucketId).then((events_window) => {
+          //events_afk = _.filter(events_afk, (e) => e.data.status == "not-afk");
+          //events_window = _.filter(events_window, (e) => e.duration > 10);
+
+          //events_afk = chunkHierarchy(events_afk, "status");
+          //events_window = chunkHierarchy(events_window, "app");
+
+          if(events_afk.length > 0 && events_window.length > 0) {
+            this.hierarchy = buildHierarchy(events_afk, events_window);
+          } else {
+            // FIXME: This should do the equivalent of "No data" when such is the case, but it doesn't.
+            console.log("HELLO");
+            this.hierarchy = {
+              "timestamp": "",
+              // TODO: If we want a 12/24h clock, this has to change
+              "duration": 0,
+              "data": {"title": "ROOT"},
+              "children": []
+            };
+          }
+        });
+      });
     },
   },
 }
