+++ src/visualizations/Sunburst.vue
@@ -1,5 +1,11 @@
 <template lang="pug">
 div.sunburst
+  div#sidebar
+    input(type="checkbox", id="togglelegend")
+    | Legend
+    br
+    div#legend(style="visibility: hidden")
+
   div#main
     div#sequence
     div#chart
@@ -12,12 +18,6 @@ div.sunburst
           div#time
           div#duration
           div#data(style="text-overflow: ellipsis; white-space: nowrap; overflow: hidden;")
-
-  div#sidebar
-    input(type="checkbox", id="togglelegend")
-    | Legend
-    br
-    div#legend(style="visibility: hidden")
 </template>
 
 <style scoped lang="scss">
@@ -30,8 +30,9 @@ div.sunburst
   margin-top: 10px;
 
   #main {
-    float: left;
     width: 750px;
+    margin-right: auto;
+    margin-left: auto;
   }
 
   #sidebar {
@@ -118,6 +119,7 @@ let aw_sunburst = {
   },
   watch: {
     "hierarchy": function() {
+      sunburst.create(this.$el);
       sunburst.update(this.$el, this.hierarchy);
     }
   }
